[ZV-219] Добавлена новая версия контракта, расширены все сервисы

+++ parameter_service/src/main/kotlin/ru/zveron/repository/CategoryRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository
 import ru.zveron.entity.Category
 
 @Repository
-interface CategoryRepository : JpaRepository<Category, Long> {
+interface CategoryRepository : JpaRepository<Category, Int> {
     fun getAllByParent(parentCategory: Category): List<Category>
 }
