+++ parameter_service/src/main/kotlin/ru/zveron/repository/ParameterFromTypeRepository.kt
@@ -7,6 +7,6 @@ import ru.zveron.entity.LotForm
 import ru.zveron.entity.ParameterFromType
 
 @Repository
-interface ParameterFromTypeRepository : JpaRepository<ParameterFromType, Long> {
+interface ParameterFromTypeRepository : JpaRepository<ParameterFromType, Int> {
     fun getAllByCategoryAndLotForm(category: Category, lotForm: LotForm): List<ParameterFromType>
 }
